+++ templates/project/Tests/SayHello/Tests.cs
@@ -2,12 +2,12 @@ using SayHello;
 
 namespace Tests.SayHello;
 
-public class Tests(Fixture f, ITestOutputHelper o) : TestClass<Fixture>(f, o)
+public class Tests(AppFixture a, ITestOutputHelper o) : TestClass<AppFixture>(a, o)
 {
     [Fact, Priority(1)]
     public async Task Invalid_User_Input()
     {
-        var (rsp, res) = await Fixture.Client.POSTAsync<Endpoint, Request, ErrorResponse>(new()
+        var (rsp, res) = await App.Client.POSTAsync<Endpoint, Request, ErrorResponse>(new()
         {
             FirstName = "x",
             LastName = "y"
@@ -21,7 +21,7 @@ public class Tests(Fixture f, ITestOutputHelper o) : TestClass<Fixture>(f, o)
     [Fact, Priority(2)]
     public async Task Valid_User_Input()
     {
-        var (rsp, res) = await Fixture.Client.POSTAsync<Endpoint, Request, Response>(new()
+        var (rsp, res) = await App.Client.POSTAsync<Endpoint, Request, Response>(new()
         {
             FirstName = "Mike",
             LastName = "Kelso"